Various updates * The singularity was not lagging enough at maximum size.  I believe I have fixed this. * The singularity will now pull turfs out of position using a structure to emulate the process.  This is laggy but produces more interesting wreckage and a more appealing spectacle. * IMPORTANT - The particle accelerator particles will now affect you in the one square directly in front of the emitters, which they previously skipped.  Do not stand there. * Mech fabricator fix - Now synchronizes properly * Corgi inventory fix - No more item duplication, some improvements * Grab fix * Admin hypospray is now its own item and therefore spawnable.  It will not appear in the game. * Adds a holy water grenade.  It will also appear in maintenance closets. * Random closet items moved out of New() and into initialize() where it belongs * When you login, your hud should tell you whether your internals are on or off. * Modifies how parrots deal with having their headsets removed and new ones added. * Clown guns will no longer spawn with a spare clip. * Work-in-progress file for the librarian's tome. * Small map fixes

+++ code/game/objects/effects/effect_system.dm
@@ -138,7 +138,8 @@ steam.start() -- spawns the effect
 					sleep(5)
 					step(steam,direction)
 				spawn(20)
-					steam.delete()
+					if(steam)
+						steam.delete()
 
 /////////////////////////////////////////////
 //SPARK SYSTEM (like steam system)
@@ -325,7 +326,8 @@ steam.start() -- spawns the effect
 					sleep(10)
 					step(smoke,direction)
 				spawn(75+rand(10,30))
-					smoke.delete()
+					if(smoke)
+						smoke.delete()
 					src.total_smoke--
 
 
@@ -425,7 +427,8 @@ steam.start() -- spawns the effect
 					sleep(10)
 					step(smoke,direction)
 				spawn(150+rand(10,30))
-					smoke.delete()
+					if(smoke)
+						smoke.delete()
 					src.total_smoke--
 
 
@@ -555,7 +558,8 @@ steam.start() -- spawns the effect
 					sleep(10)
 					step(smoke,direction)
 				spawn(150+rand(10,30))
-					smoke.delete()
+					if(smoke)
+						smoke.delete()
 					src.total_smoke--
 
 
@@ -650,7 +654,8 @@ steam.start() -- spawns the effect
 					sleep(10)
 					step(smoke,direction)
 				spawn(150+rand(10,30))
-					smoke.delete()
+					if(smoke)
+						smoke.delete()
 					src.total_smoke--
